pyCaricare/requirements.txt
Cython==0.22
leveldb==0.193
pycapnp==0.5.6
